+++ src/components/Map/index.tsx
@@ -2,7 +2,8 @@ import { EsriOceanBasemapLayer, EsriOceanReferenceLayer } from "./esriLayers"
 
 export const esriLayers = {
   EsriOceanBasemapLayer,
-  EsriOceanReferenceLayer
+  EsriOceanReferenceLayer,
 }
 
 export { BaseMap } from "./baseMap"
+export type { MapView, Props as BaseMapProps } from "./baseMap"